requirements/test.txt
-r dev.txt

coverage>=4.4
pytest>=3.3.1
pytest-cov>=2.4.0
pytest-flask>=0.14.0
